This commitment is a continuation of the effort of inetgrating the README.md, CONTRIBUTING.md, and CODE_OF_CONDUCT.md to the documentation in ReadTheDocs.
 In this version the README.md and CODE_OF_CONDUCT.md are added by a workaround where the files from the main directory (the parent directory of docs/) are called by
.md files in the dopcs directory that are called by index.rst. The workaround is that these 'bridge' files in the docs directory can be added to the index.rst
while the 'real' .md files cannot, as they are in the parent direcory.

The same approach also works for the CONTRIBUTING.md file, however it generates some warnings that I want to avoid (it still works, but I want to avoid the warnings)
the warnings are created due to cross references that MyST cannot handle well. So for this case I use 'pandoc' to directly convert the .md file in the parent directory
to a .rst file in the docs directory. I am cutious if it will work, as for instance, I am assuming that 'pandoc' is installed in the virtual machine of GitHub Actions